move schema to common
